Reorganize repo for library construction
The RBC repo will serve as the repository for where the attempt to generalize
the MPO pipeline into a library will take place.

This commit moves alot of the previous files around and adds two new classes

State is the top level class that holds the state for the MPO pipeline

Generator is the class for the data-generation stage of the pipeline.

setup_runs and run_model were placed into the generator class