Added base configuration duplication
this commit add a feature to duplicate the base configuration of a numerical model
the feature is contained within the data generation stage.

Between this feature and the models actually being run is the only model
specific portion of the MPO data generation which is actually writing the
parameters into each of the newly generated model configurations.