Improve configuration search in MPO state
this commit includes a set of functions to perform a breadth first search
through the tree of dictionaries that represent a TOML file in python.

state.get_config allows for any key within any table, including table names
themselves, to be searched for strictly by name. No parent table
information is required. returns a keyerror if the key cannot be
found.

also updated all previous references of the functions to use the new
syntax.(all in generator)