Buffed error handling
Continued to work on error handling in the configuration
writing and reading portions of the data-generation phase.

introduced MpoUnsupportedError for filetypes that have not
been implemented yet.

Future commits will take the MOM6 specific code out of the
confwriter class and make that a super class for instances
of model specific code to be written into.(or at least thats
the idea)