Remove last line of model specific code
This commit abstracts the last bit of model specifc code
from the main files that run the data generation process.

The configuration writer from here on will require that
either a developer or user write a specific configuration
writer class. An example is given in the mom6_writer file
in the newly created writers directory within data_generation.

A table, [model], was added to the toml file that will help specify which
writer to use to the data-generation phase.

Some dead code was removed that was added for the QR presentation.