change data-generation to less mpo specific interface
In this commit the format of the mpo-config.toml was changed to
a format that will support less mpo like workloads. The config file
itself was also renamed ss-config.toml.

In the ss-config.toml, the parameter and parameter info fields
have been removed. All parameters will be listed under the target
models named and listed by the user.

target models will be listed under the [execute] table where
the name of each model the user wants to be run are listed

The parameters can hold single values or arrays. if the values of a
parameter in the config file hold an array, than each value within
the array will be used to create a seperate model to be run. In this
way. Users can specify values for parameters that vary easily, and
make entirely seperate targets for parameters that dont go without
changing many other parameters like resolution.

The toml file is either found automatically or listed specifcally
in the script that runs the library created by the user.