Start Control module
This commit begins the structure of the control module
the control modules job in this first version is to interface
with the launcher that interfaces with mutliple different
workload managers.

In building the control module like this, we will have an extensible
way to interact with the various modules within the smart sim library
as each of them could need the control module at some point.
f