Develop control module
Control module beginning to be flushed out in this commit.
The ability to simulate has been added and linked with the slurm
launcher.

The interface work began in this commit as well. Control.toml
was added so that the control package could be used entirely by
itself.

State was modified so that another configuration file can be passed
to state. In this way, the state configuration file will be checked
before any sub-config file such as "control.toml". In doing this, we
effectively make the ss-config.toml the overriding configuration file
for each of the modules. the user will never need to interface with
any of the individual module files unless they are using the module
independant of the entire library.

Something of note is that State will need to be an internal module
moving forward. If a user doesnt init a State then we must do so.
To do this, each module will be responsible for the creation of a
state if one is not passed to them.