Add parent class to ss modules
SSModule is now the parent class that retieves confiurations
for the submodules in the preferred hierarchy.

This replaces the BFS that was used to search for configs in the
first instance as a few headaches came about when trying to search
for configs with different parent tables but the same key name. The
full path is now required, but the need to call from state is abstracted
through the introduction of this new parent class.

Overall this just cleans up how we will introduce new modules and how
the modules will interface with the global configuration file.