Change how target configuration files are found
Before, the user of Smartsim was required to place
all the necessary files for a given simulation in
an array within the simulation.toml such as

```toml
[model]
configs = ["config1", "config2"....]
```

This was to high of a barrier to usage for most cases where
there are a number of simlation configuration files that dont
need any files to be changed but must be present in a model
directory for that model to be simulated.

the current approach allows for a user to either specify the
file by name or place all necessary files within a directory
and just list that one directory.

the Generator class copies all files and directories listed,
and then the modelwriter performs a walk to find, read, and write
to all of them.