Update Controller for new Launcher
The methodology behind the data generation has changed
and thereore the controller needs to change as well.

This commit largely reverts the previous implementation as the
launcher that was being used has been replaced.

This preliminary commit introduces the two new ways that
a model can be run. first, just simply on a machine that has
mpi or open_mpi configured. Second, through three different
launchers
  - Urika
  - slrum
  - PBS

This commit only implements the first way of running models,
but future commits will flesh out the implementation.

This code as it stands runs parallel LAMMPS codes.