Update and extend logging functionality
Second bit of progress made on the logging functionality
of SS. This commit introduces a new dependency "coloredlogs"
and uses that library to write informative messages to the
user in a variety of colors.

The logger is kept as a part of state currently which will
make it easy to load and unload logs, but we may want to
make it a globally defined variable so that all modules within
the project including error can use it. look for that in a
future commit