Add SlurmLauncher into controller
This commit adds the SlurmLauncher from the poseidon launcher
into the codebase for SmartSim.

there are three ways a user can specify arguments for the running
of simulations.

1) under the execute table

e.g.
```toml
[execute]
executable = "MOM6"
```

2) Under the target table

e.g.
```toml
[execute]
  [execute.some_target]
  run_args = "-np 6"
```

3) in the initialization of the controller

```python
control = Controller(run_args="-np 6", executable="MOM6")
```

These are subject to change but are the current methodology for the
control of the SmartSim library.