Stress test MOM6 model creation
This commit fixes a few bugs in the controller implmentation
but it is more so meant to be a lock for the stress test of
the current smartsim library as a whole.

two targets were run in the experiment with a total of 147 nodes
allocated at once. All simulation data was recieved correctly and
the models logging and errors were captured.

Boom goes the dynamite