Add CP2K example, h2o simulation on GPU
this commit adds in the CP2K model to the folder of examples
that already includes LAMMPS and MOM6. The simulation is an
h2o simulation that was taken from the tests repository within
the CP2K repo.

The simulation.toml in the example specifies 4 different configurations
of the h2o simuatlion to be run within a single target and with each
model running on a single GPU.

On Kepler K20's, these four simulation models run in under a minute.