Hide internal methods and fields
Some of the methods for the generator and the controller
classes were exposed to the use of the smartsim library.
this was not the intended design and was part of the
rapid prototyping that this project was built on top of.

Now that the project is more mature and is being used, it is
imporant that good encapsulation practices are used so that
the user is not confused by what is and what is not an internal
method or field.