Add poll and finished to controller interface
Add two new methods for uses to see if their models have finished
both manually and programatically.

The commit also introduces the Job class which aides the controller
in keeping track of the various information useful to the user when
models have been submitted.

This commit only works if the patch has been applied to the
poseidon launcher.