Add ModelWriter tests
the modelwriter will have improvements coming in the next few days
and we want to make sure that the new features do not effect the
existing functionality.

this commit introduces a number of new files most of which are config
files that will be used for testing.

the test_config dir will have multiple sets of test varying by difficulty
and purpose. the first set of tests introduced are the "easy" tests where
placeholders, numbers, and strings are replaced within a config file.