Make regex for modelwriter tag chooseable
the user can now set the tag(single character) or regex(full regular expression)
for the modelwriter.

interface method in the generator allow for the tag to be set before any of the
model are generated. this allows for any type of configuration file to be written
as now there are no restrictions on what a model file can have within it.