Refactor Errors
The errors didnt utilize the logging system that was built into
the state and smartsim modules. this commit changes that so that
all errors now use the logging format and are outputted using the
same logger stored in state.

read_config was turned into an internal state method.

other minor changes for the refactor were included.