First pass at the processor module
The interface for the processor module will allow for a
number of different processors to be defined. These processors
will slowly be added over time to support new file types that
come from simulation model data.

The first processor being implemented is the netCDFProcessor that
will be able to handle the data coming out of the MOM6 model.

the idea behind the processor story is two fold

either:

 1) a user wants a reduced tensorflow style dataset from the
    output of their simulation models

or

 2) The user can offload all the paths quickly and easily to a
    variable through one call to the processor/state.