Define processor strategy
the processing stage of any machine learning experiment is complex
and largely depends on the data that the experiment is using.

For this reason, we dont want to box users into a single way of
preprocessing, rather, we want to do the opposite but provide
the convience as if it was a pre-written module.

To accomplish this, a processor base class is estabilished
that provides methods to interact with State, targets, and models.
The class provides two functions that retrieve the locations
of the model data. The class also allows for an arbitrary number
of class attributes to be defined by the user.

An example of such a processor implementation is included in the
examples directory where the MPO research processing files were
modified to work with the processor class.