Save and load pickled targets
This commit replaces the previous method of loading targets
with a method that leverages pythons pickle module.

Also added is a new state interface method "save" that allows
the user to save all targets to an experiment.

Currently the user cannot load targets from another experiment
to run them, but that functionality could be something we target,
pun intended, in the future.

Also included in this commit:
      - some changes to error messages
      - small bug fix in controller module with run_with_command
      - some TODOs