Add  methods to the controller interface
The Controller interface needed methods for querying the jobs
within the controller itself. These methods add that functionality
and allow for users to obtain that list of jobs as well as
query which nodes they are running on.

this commit also promotes model.py out of generation into the base
directory.