Add orchestrator class
the orcestrator class will handle the launching of the KeyDB and the
various connections that need to be handled. The orc is a SmartSimModule
and has access to the entire suite of configurations

currently everything is hardwired to run through a specific node, but
the idea is that hopefully soon the orchestrator will be able to
facilitate the connections between nodes and simulation models.

Soon, the orc will also handle the launching of nodes.

an environment variable was set so that the DB can find the
configuration file