Decrease data transfer for training loop by 4x
the ``send_big_data`` method is introduced in this commit. The method
uses a spliting feature to split the bytes into 4 seperate messages that
each get stored in the orchestrator at the key + : + num_split.

A header is used to indicate that the message is a multipart message.
A call to get_data now looks to see if that key:value is a multipart
message and routes to _get_multipart if it is.

this method is shown to decrease the time spent in data transfer by
up to 4x. See results in test_online_training for results.