Added in changes to work with empty targets.
Now, default targets are created and used, including a 'default_target'.
_create_model in generator has been changed to allow for 'empty' sets of
parameters; before, it calls the permutation strategy, which _may_
result in 'empty' models.  Now, if Target.get_target_params returns an
empty set of lists, it skips adding models.

A second test for this scenario has been created and added in this
commit, as well.