Added in more functions to the logger.
Can now set the 'global' logging level, or an individual logger level;
the user can also set every logger call to be debug, such that any call
to a logger is printed to the stream (either a file or stdout).

Also added the function "log_to_file", which allows logging output to
also be sent to a file.  Note that this doesn't stop the logger from
also sending to the stdout stream.