Add online inference tests
The profiling of the online training and inference loops are complete
with the addition of the online inference tests

also added is a mechanism by which a client can block until a value
is stored at a specific key. The rate at which the client polls is
also controllable by the user.