Make SmartSim use getcwd instead of SS_HOME
This has been a long-standing issue in SmartSim, the output
of the generator and associated functions outputs to SMARTSIMHOME
instead of the working directory which would be what the user expects.

This also remove the requirement that the user source the SmartSim
setup script each and everytime the library is to be used.

Tests and some examples were updated