Vendor Pos Launcher
Moving forward it will be best to maintain the poseidon launcher
within the smartsim repository. This way all of the tests can
be included and various functionality that is needed now and
down the road can be directly placed into the library.

there are still many things that need to be completed to finish
the vendoring process.
  - tests (shell is started)
  - PBS integration (started)
  - clean up documentation
  - move errors to error folder