Simplify Build Process
After the vendor of the launcher, the last step to make the dev build
a bit easier was to make the setup_env.sh handle the installation
of KeyDB. If installed, the existing installtion will be used. If
KeyDB has been downloaded in smartsim dir then it will add it to the
path, otherwise it will download, make it, and add it to the path.