Explicit kwargs and updates test_generator
This PR makes arguments that are required for the controller
and generator explicit required constructor parameters instead
of folding these parameters into kwargs. The purpose of this
is to make these user facing objects more usable. A test was
also added to make sure that the generator can create experiment
directory structures without the specification of model files.