Start Launcher Refactor
The launcher as it stood had too much code for what we needed
and covered a massive amount of functionality that we did not
want to maintain in a vendored version of the launcher.

The main interface is now running though allocations.
Users will call get_alloc() followed by any number of
run_on_alloc()