Run_command fix
This commit should fix the bug in the controller where the run_command
could be listed twice. The run_command argument is only used now when
the local launcher is being used. Otherwise when using the Slurm
launcher, all pieces of the run_command that arent srun such as "python"
should live in the executable as in "python exe.py"