Add SlurmLauncher.validate()
Added in this commit are a number of methods that make the SlurmLauncher
much better at validating and retieving information about jobs. tests
were added that ensure that SlurmLauncher.validate works as expected as
well as the information retrieval methods themselves.

The controller now has a dedicated function determining if all of the
user listed jobs have enough space on the given partition to all run
at once.