First pass at fixing local launcher
The local launcher had been broken ever since the single
allocation work and because of that, the LAMMPS example was
also broken.

This is an interim commit that introduces a new local launcher
class and database clusters of arbitrary size. These two changes
were naturally made together as they changed the same part of the
code base. Other minor refactors are detailed below.

other
  - when generated, node and orchestrator live in experiment
  - single database nodes can now be created, if cluster_size < 3
    the dbnodes are started in standalone mode.
  - helpers.py introduced into utils
  - Entity now can reset paths within the run_settings.
  - cmd is not kept within run_settings, now a entity class variable