Wrap up local launcher; Add warnings to controller
the local launcher currently launches jobs in serial and does
not support launching an orchestrator in parallel with jobs.

in the future we plan to make it so that the jobs can be launched
in parallel and with an orchestrator. function stubs are included
in the local launcher that should make that straightforward when
we get to it.

Warnings were added to the controller to make sure that users do
not call those functions when launching locally.