Job manager
Implemented the job manager class which tracks the various jobs
by entity type and stores them in a number dictionaries. Introduction
of dunder methods in the class made the code change to controller
minimal.

entity type is now a initialization value of the SmartSimEntity class.

fixed small bug for orchestrator cluster output