Add profiling test of cluster/non-clustered python
the python client profiling tests were updated to be
able to test both the clustered and non-clustered versions
of the client.

Error messages for debugging where added to the controller
that makes the launching process a bit more tractable

finished was given an extra argument that allows it to ignore
the database nodes.