Update Env Var connection system; add heat-sim; update python client
The connection system used to use SSDATAOUT which has now been replaced
with simply using SSNAME to prefix keys being sent to the database. The
python client was updated to reflect this change.

The heat simulation example was added to showcase the c++ client. This
example currently online runs on cicero