Initial implementation of command center
The command center is a method by which smartsim can
be run on the compute nodes of a cluster or super.
The approach uses a ZMQ server to communicate launcher
commands back to a head node where slurm or other
workload managers are installed.

Various interface methods have been added to experiment
in order to ease the use of such methods. This plays
extremely well for using the Urika data analytics stack
alongside SmartSim.