Add features to get and add allocation to experiment
The controller and launcher were modified such that
existing allocations can be run on consecutively and
user allocations can be accepted into the controller.

The biggest change is that the launcher class now
manages all allocations and the controller simply
interfaces with the launcher for allocation manipulation.

Methods to obtain and accept allocations from the user
were also put into the controller

Squashed commit of the following:

commit 0355640456cf4096da2b32ccea35243a64e6b604
Author: Sam Partee <partees21@gmail.com>
Date:   Mon Mar 23 17:01:01 2020 -0500

    Add experiment interface methods for allocations

    Methods for getting and adding existing allocations
    have now been exposed through experiment.

commit eeb666ff85584fc967ac0d01e682c55b50b6d027
Author: Sam Partee <partees21@gmail.com>
Date:   Fri Mar 20 18:10:26 2020 -0500

    Initial work on running on existing allocs

    The controller and launcher were modified such that
    existing allocations can be run on consecutively.

    The biggest change is that the launcher class now
    manages all allocations and the controller simply
    interfaces with the launcher for allocation manipulation.

    Methods to obtain and accept allocations from the user
    were also put into the controller