Vertical Scaling of Orchestrator
Add database per node functionality

The ability to run multiple databases per node was introduced
in this feature. Through experiment, users can now create
an orchestrator with dpn > 1 which will launch a --multi-prog
slurm job.

Secondly, an interface to connect a jupyter notebook
or smartsim script to the orchestrator was added to the
experiment interface.

The database cluster configuration was also updated to the latest
preferred settings.

Squashed commit of the following:

commit 5c02e1f6029f6ac6d6fe4b384fd777cbdcc921ea
Author: Sam Partee <partees21@gmail.com>
Date:   Mon Apr 13 14:31:10 2020 -0500

    Update to database configuration file

    The following settings were changed in the KeyDB config file

     - activerehashing -> No
     - require-cluster-full-coverage -> No
     - server-threads -> 8

commit 59852459619e98bc67cc52673438e5b3c08271a5
Author: Sam Partee <partees21@gmail.com>
Date:   Thu Apr 9 14:27:12 2020 -0500

    Add database per node functionality

    The ability to run multiple databases per node was introduced
    in this feature. Through experiment, users can now create
    an orchestrator with dpn > 1 which will launch a --multi-prog
    slurm job.

    Secondly, an interface to connect a jupyter notebook
    or smartsim script to the orchestrator was added to the
    experiment interface.