Update LAMMPS examples
This updates the local lammps example to
use the slurm launcher.

As well, this commit changes the configuration
of the example to one that doesnt require
any external lammps packages

lastly, a small bug in the creation of Slurm
steps in slurmStep.py was addressed.

Squashed commit of the following:

commit 8b0059ad06dc9a8cc10c50ad579b756d49277842
Author: Sam Partee <partees21@gmail.com>
Date:   Tue Jul 14 16:19:31 2020 -0500

    Update local LAMMPS example to slurm

    This updates the local lammps example to
    use the slurm launcher.

    As well, this commit changes the configuration
    of the example to one that doesnt require
    any external lammps packages

    lastly, a small bug in the creation of Slurm
    steps in slurmStep.py was addressed.