Fix experiment.start bug; address test failures
A bug in experiment.start was fixed in a
previous commit, this however broke the
ability to reconnect to an existing
orchestrator because those tests relied
on having the bug in place.

Users are now required to call the reconnect
function and then supply that orchestrator
object returned by reconnect_orchestrator
to the experiment.start function.

Changes to reflect this have been made in the
test suite so that now all tests pass