Address thread sync with locks
The job manager and task manager have various
places that needed to be under a locking data
structure such that the threads will not run
over data in either manager.

This implementation still requires more testing
but should be stable enough for most experimentation
with SmartSim

Also addressed a small bug in how slurm arguments
are passed.