Catch errors between Ppoen and WLM
Sometimes an error can occur between the time when
the command is sent to the OS and before Slurm
registers the job.

An example of this is an improperly formed slurm
command. e.g. srun -not-an-arg 2

We now catch such errors as long as they occur within
the time frame described above. We should be able to
provide a better user experience when throwing
errors on launch because of this.

tested on horizon/slurm. 83% coverage.