Add Cobalt Launcher Support
Cobalt launcher support has been added for
launching in interactive allocations obtained
with qsub as well as batch allocations.

Largely, the launcher implementation mirrors
PBS and should at some point be condensed possibly
as a subclass of the PBS implementation.

All tests pass on Theta.

Also added are two important functions for launching
on congested systems
 - improved TM.remove_task for killing child processes
 - controller.orchestrator_launch_wait

The latter ensures that even if the orchestrator is queued
in the batch system, SmartSim will wait for a running
status before looking for the RedisIP files.

The Cobalt Launcher does not currently support node
list aquisition.