Add editable configuration and build
Add a configuration layer so that future
additions to configuration do not require refactors.

mostly this was to support existing installations
of Redis, RedisAI, and RedisIP which can now
all be installed by the user.

Also edited was the pip install process and the
library builds as a wheel and a dist