Modify launch_database_cluster to run in CIME
To ensure that the vanilla OpenMPI launcher is used to launch the
database, the database cluster launcher script needs to be wrapped
within a bash script. This also comments out much of the code
within the launching script so that the database will live for the
duration of the allocation.