Add is_active() to check that database is active
If the user stops the databse, the port and host
fields would still be populated. With is_active()
we can now check that the database is active
before populating addresses.

Made get_address() a public method.

Added an explicit error message in get_address()
if database address can't be found.