Address local launcher for Ray
The local launcher for Ray was broken
on Mac because of how Ray does IP lookups
for local addresses.

The local launcher functionality has been
taken out and the slurm and PBS are the
only supported launchers now.

Minor cleanup on RayCluster classes as
well. Changed inheiriance structure from
model to SmartSimEntity