Remove TOML configuration option
The TOML configuration option seemed to
confuse users and we really want most users
to use the default settings anyway. The purpose
of the configuration was to ensure each site
could setup SmartSim in the way they wanted with
a global configuration.

This commit replaces that with simple environment
variables that provide the same functionality.

Site's wishing to have a site-wide configuration
can now set this environment variables upon
loading the SmartSim module