Clean up output and handle CUDNN
SmartSim #122 brought up a user issue with
CUDNN environment variables. To address this
the cli now sets both the caffe and torch env
vars in the RedisAI build

as well, the logging module is now used within
the CLI for better output during the build process.
this replaces our manual Warning_ class