Introduce _core module (#122)
Introduce _core module (#122)

Given the growth of the library there
was a need to separate out the user
facing functionality from the backend
of which users are not supposed to import

the structure now includes a _core dir
which houses modules which are not using
facing.

the following have been moved to _core
  - control
  - generation
  - launcher
  - bin and lib
  - utils

Remove TOML configuration option

The TOML configuration option seemed to
confuse users and we really want most users
to use the default settings anyway. The purpose
of the configuration was to ensure each site
could setup SmartSim in the way they wanted with
a global configuration.

This commit replaces that with simple environment
variables that provide the same functionality.

Site's wishing to have a site-wide configuration
can now set this environment variables upon
loading the SmartSim module

Mirgrate the log and slurm modules

log and slurm modules should have always been
at the top level as they are expected to be
imported by the user.

we may move to a ``wlm`` module in the future
that contains user-facing functions for certain
workload managers.

Rename constants to status

instead of:

 > from smartsim import constants
 > constants.STATUS_COMPLETED

users should now do

 > from smartsim import status
 > status.STATUS_COMPLETED

* CLI and build improvements (#94)

the cli now has it's own module
within _core/_cli. Each function is
now split out into it's own file which
will make it much easier to add new additions
in the future.

The build process controlled by setup.py and
the ``smart`` cli used to be governed by two
different build methods mostly contained in
bash scripts.

these approaches have been unified into the
_core/_install directory which replaces the
bash scripts we used to use and works for both
the cli and setup.py.

There are a few non-obvious benefits to this.
 - all versions contained in one place
 - versions can be changed through env vars
 - build process can be changed through env vars
 - multiple RAI versions can now be supported
 - multiple DL framework versions can now be supported