Docker Documentation Builds (#133)
This PR adds the ability to build the documentation
inside a docker build so that users don't have to worry
about the specific build steps. The build exports the
docs folder into the current clone so that the dev
 doesn't have to do it themselves. build with
``make docks``. The manual documentation
build is still available with make docs.

There is also a new make tutorials that builds
the tutorials into a developer docker with the
current clone of SmartSim. This will eventually
have a prod docker build too whch downloads the
pinned version of smartsim for the latest release.

[ committed by @spartee ]
[ reviewed by @al-rigazzi ]