Controller.get_job_data and log _launch_summary
The scaling tests used the old Experiment.summary()
method to obtain job data. Since we decided to refactor
that to remove the pandas requirement, get_job_data
was introduced to obtain data from the Controller
instance.

We hide job data from the user intentionally, which
is why the Controller is a private variable in the
Experiment, but this method will be expanded upon
later to provide more job data (with filters) once
we have a reliable, persistent accross experiment run,
database.

Experiment._launch_summary() was also refactor to use
the logger instead prints and handle the co-located
Model case.