Address process killing on interrupt
This commit introduces the interrupt strategy
needed to handle jobs and tasks during a keyboard
interrupt.

follow ups will include the parameter to kill
all tasks on interrupt. this commit is primarily
geared towards informing users of colocated tasks
that may still be running when an interrupt is
triggered.

A few changes were also made to the colocated launch
script as it seems psutil does not support cpu pinning
on MacOS