Support for site-provided dependencies (#182)
Support for site-provided dependencies

Include an optional argument`--only_python_packages` to
to `smart build`. This will also issue a warning for a
case where a user might accidentally omit the device flag (which
defaults to cpu) and not install the GPU variant of torch.

The config can now also detect whether dependencies are available
elsewhere on the platform. If so, those libraries and executables are
not built installed within the user's conda environment.

Reviewed by @Spartee