Singularity docs in SmartSim API (#205)
Add Singularity docs to SmartSim API docs.

I was on the fence about exposing the base class, `Container`. It has no use for a user, but it is the  type referenced in arguments, e.g. 

```python
  def create_run_settings(self, ..., container: Container, ...):
    ...
```

I opted for leaving `Container` docs out in this PR, but am open to changing that if the reviewer disagrees. This decision can also be revisited in the future, if users are finding the current form confusing.

[Commited by @ben-albrecht]
[Reviewed by @ashao]