Fix singularity local launcher (#210)
* Fix singularity local launcher

When not using Apptainer, the singularity tests were failing on
the local launcher because the SmartSim-specified run directories
were not being mounted. To solve this, when the launch command
is constructed, the working directory is passed to the container's
command constructor and automatically bound. For advanced users,
the working directory can be specified when the container object
is created in which case that directory takes precedence.