SmartRedis Integration Guide (#214)
Based on feedback from users, the SmartRedis documentation needed
a guide for application developers to integrate into an existing
application. A new page to the documentation has been added to describe
the common tasks that typically need to be done. Basic code in each of the
four languages (C/C++/Python/Fortran) are included for the general
guidelines with a full Fortran-specific example.

[ commited by @ashao ]
[ reviewed by @juliaputko and @billschereriii ]