Refactor mpiexec/mpirun steps and settings
mpiexec, mpirun, and orterun are all synonyms in the modern MPI
standards. The mpirunSettings and mpirunStep have been refactored
to the more generic mpiSettings and mpiStep. Within these modules
the classes associated with mpiexec, mpirun, and orterun all
inherit from mpiexec (which historically was the first executable
method that was defined in the MPI standard).

The one exception to this is that the slurm workload manager
provides a wrapper to mpiexec which wraps a subset of the
supported options to mpiexec so that they can be passed to Slurm.
In general, this is a very limited set of options and so we choose
not to support it. Instead, we check to see if mpiexec is actually
mpiexec.slurm and throw an UnsupportedError.