Domain Socket Support for co-located databases (#246)
Co-located databases (i.e. a database deployed on the same node as the 
application) can now be configured to accept connections over Unix Domain
sockets instead of the loopback interface. This yields performance benefits
in the form of lower latency communication. Note that this PR depends on 
developments that were added to SmartRedis.

[ committed by @ashao ]
[ reviewed by @MattToast ]