Update installation instructions (#262)
The changes here describe how to install SmartSim on a wide
variety of platforms. Some of this material was previously available, but
in general should be easier to follow for users. In particular, this PR 
includes instructions for how to build SmartSim with the GPU-enabled
ML backends and elucidates the process for using 'site-install' of
SmartSim.
[ committed by @ashao ]
[ reviewed by @amandarichardsonn @al-rigazzi @MattToast ]