Clean up environment before each test (#270)
This PR fixes the behavior of the test suite with respect to SmartSim-specific environment variables. Such variables are now removed from the environment prior to each test. A warning for Slurm's environment variable handling was added, to make users aware when a variable will not be exported as expected with `--export`.

[ committed by @al-rigazzi ]
[ reviewed by @ankona @MattToast ]