Updates Fortran documentation for error codes
The examples in the Fortran documentation examples used a mixture
of subroutines and functions. The former was deprecated when
error handling was implemented to allow the clients to bubble up
a return code that Fortran applications can check. This has now
been updated to allow others to use the documents as an accurate
jumping off point.